tests: Move `GET /crates/:name/reverse_dependencies` tests to `krate::reverse_dependencies` module
